chess-app/soap-adapter/adaptor/requirements.txt
Flask==2.1.2
grpcio==1.43.0
grpcio-tools==1.43.0
protobuf==3.20.3
lxml==4.6.3
werkzeug==2.0.3